refactor: 재사용 가능한 UI 컴포넌트를  component  폴더로 분리
