+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Dive"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".screen.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Dive">
