+++ app/src/main/java/com/sopt/dive/component/TextField.kt
@@ -1,0 +1,51 @@
+package com.sopt.dive.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.LocalTextStyle
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun LabeledTextField(
+    label: String,
+    placeholder: String,
+    text: String,
+    onValueChange: (String) -> Unit,
+    isPassword: Boolean = false
+) {
+    Column(modifier = Modifier.fillMaxWidth()) {
+        Text(
+            text = label,
+            fontSize = 20.sp,
+            color = Color.Black
+        )
+        TextField(
+            value = text,
+            onValueChange = onValueChange,
+            placeholder = { Text(placeholder, color = Color.Gray, fontSize = 15.sp)},
+            singleLine = true,
+            modifier = Modifier
+                .fillMaxWidth(),
+            textStyle = LocalTextStyle.current.copy(
+                fontSize = 15.sp,
+                color = Color.Black
+            ),
+            visualTransformation = if (isPassword) PasswordVisualTransformation()
+            else VisualTransformation.None,
+            colors = TextFieldDefaults.colors(
+                focusedContainerColor = Color.Transparent,
+                unfocusedContainerColor = Color.Transparent,
+                focusedIndicatorColor = Color.Black,
+                unfocusedIndicatorColor = Color.Black
+            )
+        )
+    }
+}
