+++ app/src/main/java/com/sopt/dive/component/Title.kt
@@ -1,0 +1,21 @@
+package com.sopt.dive.component
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun Title(text: String) {
+    Text(
+        text = text,
+        style = MaterialTheme.typography.titleLarge,
+        color = Color.Black,
+        textAlign = TextAlign.Center,
+        fontSize = 30.sp,
+        fontWeight = FontWeight.Bold
+    )
+}
